requirements.txt
python-telegram-bot==20.7
litellm==1.28.14
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0 